requirements.txt
flask>=3.0.0
flask_sqlalchemy>=3.0.0
sqlalchemy>=2.0.0
flask_login>=0.5.0
werkzeug>=1.0.0
python-dotenv>=1.0.0